tests/wpt/tests/html/semantics/scripting-1/the-script-element/script-onload-insertion-point.html

<!doctype html>
<meta charset=utf-8>
<title>Test that the insertion point is defined in the load event of a parser-inserted script.</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script>
  var t = async_test("");
  var writeDone = t.step_func_done(function(text) {
    assert_equals(text, "Some text");
  });
</script>
<iframe src="support/script-onload-insertion-point-helper.html"></iframe>
